drm/i915/dg2: Do not use phy E
PORT_TC1 is still not being initialized - that is the port that uses phy
E.  However the intel_phy_is_snps() reports that phy as being present,
which causes warnings about unclaimed access to the PHY_MISC register.
Even with some basic wiring up for that port, it still gives the error:

	i915 0000:03:00.0: [drm] *ERROR* SNPS PHY E failed to calibrate after 25ms.

So remove it for now.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>